refactor(wsLink): Split createWSClient from wsLink
