+++ packages/client/src/links/wsLink/createWSClient.ts
@@ -1,42 +1,24 @@
-import type { Observer, UnsubscribeFn } from '@trpc/server/observable';
-import { behaviorSubject, observable } from '@trpc/server/observable';
-import type { TRPCConnectionParamsMessage } from '@trpc/server/rpc';
+import type { AnyRouter, inferRouterError, ProcedureType } from '@trpc/server';
+import {
+  behaviorSubject,
+  type Observer,
+  type UnsubscribeFn,
+} from '@trpc/server/observable';
 import type {
-  AnyRouter,
-  inferClientTypes,
-  inferRouterError,
-  ProcedureType,
   TRPCClientIncomingMessage,
   TRPCClientIncomingRequest,
   TRPCClientOutgoingMessage,
+  TRPCConnectionParamsMessage,
   TRPCRequestMessage,
   TRPCResponseMessage,
-} from '@trpc/server/unstable-core-do-not-import';
-import { transformResult } from '@trpc/server/unstable-core-do-not-import';
-import { TRPCClientError } from '../TRPCClientError';
-import type { TransformerOptions } from '../unstable-internals';
-import { getTransformer } from '../unstable-internals';
-import type { TRPCConnectionState } from './internals/subscriptions';
+} from '@trpc/server/unstable-core-do-not-import/rpc';
+import { TRPCClientError } from '../../TRPCClientError';
+import type { TRPCConnectionState } from '../internals/subscriptions';
 import {
   resultOf,
   type UrlOptionsWithConnectionParams,
-} from './internals/urlWithConnectionParams';
-import type { Operation, TRPCLink } from './types';
-
-const run = <TResult>(fn: () => TResult): TResult => fn();
-
-type WSCallbackResult<TRouter extends AnyRouter, TOutput> = TRPCResponseMessage<
-  TOutput,
-  inferRouterError<TRouter>
->;
-
-type WSCallbackObserver<TRouter extends AnyRouter, TOutput> = Observer<
-  WSCallbackResult<TRouter, TOutput>,
-  TRPCClientError<TRouter>
->;
-
-const exponentialBackoff = (attemptIndex: number) =>
-  attemptIndex === 0 ? 0 : Math.min(1000 * 2 ** attemptIndex, 30000);
+} from '../internals/urlWithConnectionParams';
+import type { Operation } from '../types';
 
 export interface WebSocketClientOptions extends UrlOptionsWithConnectionParams {
   /**
@@ -47,7 +29,7 @@ export interface WebSocketClientOptions extends UrlOptionsWithConnectionParams {
    * The number of milliseconds before a reconnect is attempted.
    * @default {@link exponentialBackoff}
    */
-  retryDelayMs?: typeof exponentialBackoff;
+  retryDelayMs?: (attemptIndex: number) => number;
   /**
    * Triggered when a WebSocket connection is established
    */
@@ -96,18 +78,42 @@ export interface WebSocketClientOptions extends UrlOptionsWithConnectionParams {
   };
 }
 
+type WSCallbackResult<TRouter extends AnyRouter, TOutput> = TRPCResponseMessage<
+  TOutput,
+  inferRouterError<TRouter>
+>;
+
+type WSCallbackObserver<TRouter extends AnyRouter, TOutput> = Observer<
+  WSCallbackResult<TRouter, TOutput>,
+  TRPCClientError<TRouter>
+>;
+
+class TRPCWebSocketClosedError extends Error {
+  constructor(opts?: { cause?: unknown; message?: string }) {
+    super(
+      opts?.message ?? 'WebSocket closed',
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore https://github.com/tc39/proposal-error-cause
+      {
+        cause: opts?.cause,
+      },
+    );
+    this.name = 'TRPCWebSocketClosedError';
+    Object.setPrototypeOf(this, TRPCWebSocketClosedError.prototype);
+  }
+}
+
+const run = <TResult>(fn: () => TResult): TResult => fn();
+
+const exponentialBackoff = (attemptIndex: number) =>
+  attemptIndex === 0 ? 0 : Math.min(1000 * 2 ** attemptIndex, 30000);
+
 type LazyOptions = Required<NonNullable<WebSocketClientOptions['lazy']>>;
 const lazyDefaults: LazyOptions = {
   enabled: false,
   closeMs: 0,
 };
 
-/**
- * @see https://trpc.io/docs/v11/client/links/wsLink
- * @deprecated
- * 🙋‍♂️ **Contributors needed** to continue supporting WebSockets!
- * See https://github.com/trpc/trpc/issues/6109
- */
 export function createWSClient(opts: WebSocketClientOptions) {
   const {
     WebSocket: WebSocketImpl = WebSocket,
@@ -635,103 +641,4 @@ export function createWSClient(opts: WebSocketClientOptions) {
   };
 }
 
-/**
- * @see https://trpc.io/docs/v11/client/links/wsLink
- * @deprecated
- * 🙋‍♂️ **Contributors needed** to continue supporting WebSockets!
- * See https://github.com/trpc/trpc/issues/6109
- */
 export type TRPCWebSocketClient = ReturnType<typeof createWSClient>;
-
-/**
- * @see https://trpc.io/docs/v11/client/links/wsLink
- * @deprecated
- * 🙋‍♂️ **Contributors needed** to continue supporting WebSockets!
- * See https://github.com/trpc/trpc/issues/6109
- */
-export type WebSocketLinkOptions<TRouter extends AnyRouter> = {
-  client: TRPCWebSocketClient;
-} & TransformerOptions<inferClientTypes<TRouter>>;
-class TRPCWebSocketClosedError extends Error {
-  constructor(opts?: { cause?: unknown; message?: string }) {
-    super(
-      opts?.message ?? 'WebSocket closed',
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore https://github.com/tc39/proposal-error-cause
-      {
-        cause: opts?.cause,
-      },
-    );
-    this.name = 'TRPCWebSocketClosedError';
-    Object.setPrototypeOf(this, TRPCWebSocketClosedError.prototype);
-  }
-}
-
-/**
- * @see https://trpc.io/docs/v11/client/links/wsLink
- * @deprecated
- * 🙋‍♂️ **Contributors needed** to continue supporting WebSockets!
- * See https://github.com/trpc/trpc/issues/6109
- */
-export function wsLink<TRouter extends AnyRouter>(
-  opts: WebSocketLinkOptions<TRouter>,
-): TRPCLink<TRouter> {
-  const transformer = getTransformer(opts.transformer);
-  return () => {
-    const { client } = opts;
-    return ({ op }) => {
-      return observable((observer) => {
-        const { type, path, id, context } = op;
-
-        const input = transformer.input.serialize(op.input);
-
-        const connState =
-          type === 'subscription'
-            ? client.connectionState.subscribe({
-                next(result) {
-                  observer.next({
-                    result,
-                    context,
-                  });
-                },
-              })
-            : null;
-        const unsubscribeRequest = client.request({
-          op: { type, path, input, id, context, signal: null },
-          callbacks: {
-            error(err) {
-              observer.error(err);
-              unsubscribeRequest();
-            },
-            complete() {
-              observer.complete();
-            },
-            next(event) {
-              const transformed = transformResult(event, transformer.output);
-
-              if (!transformed.ok) {
-                observer.error(TRPCClientError.from(transformed.error));
-                return;
-              }
-              observer.next({
-                result: transformed.result,
-              });
-
-              if (op.type !== 'subscription') {
-                // if it isn't a subscription we don't care about next response
-
-                unsubscribeRequest();
-                observer.complete();
-              }
-            },
-          },
-          lastEventId: undefined,
-        });
-        return () => {
-          unsubscribeRequest();
-          connState?.unsubscribe();
-        };
-      });
-    };
-  };
-}
